*** empty log message ***

+++ src/elements/DEM_ellip3d/plot/sortParaIJK/sortParaIJK.cpp
@@ -84,7 +84,7 @@ public:
   Vec    velocity;
   double pressure;
   double temperature;
-  int    mask;
+  double mask;
   Vec    penalForce;
   Vec    pressureForce;
